# typescript-fetch/models/Student.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online School
 * Online School Application is where students and teachers meet together
 *
 * The version of the OpenAPI document: 0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    StudentAllOf,
    StudentAllOfFromJSON,
    StudentAllOfFromJSONTyped,
    StudentAllOfToJSON,
    StudentBase,
    StudentBaseFromJSON,
    StudentBaseFromJSONTyped,
    StudentBaseToJSON,
    StudentStatus,
    StudentStatusFromJSON,
    StudentStatusFromJSONTyped,
    StudentStatusToJSON,
} from './';

/**
 * 
 * @export
 * @interface Student
 */
export interface Student {
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Student
     */
    name: string;
    /**
     * 
     * @type {StudentStatus}
     * @memberof Student
     */
    status: StudentStatus;
}

export function StudentFromJSON(json: any): Student {
    return StudentFromJSONTyped(json, false);
}

export function StudentFromJSONTyped(json: any, ignoreDiscriminator: boolean): Student {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': json['name'],
        'status': StudentStatusFromJSON(json['status']),
    };
}

export function StudentToJSON(value?: Student | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'status': StudentStatusToJSON(value.status),
    };
}