# typescript-fetch/models/StudentBase.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online School
 * Online School Application is where students and teachers meet together
 *
 * The version of the OpenAPI document: 0.1.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    StudentStatus,
    StudentStatusFromJSON,
    StudentStatusFromJSONTyped,
    StudentStatusToJSON,
} from './';

/**
 * 
 * @export
 * @interface StudentBase
 */
export interface StudentBase {
    /**
     * 
     * @type {string}
     * @memberof StudentBase
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof StudentBase
     */
    name: string;
    /**
     * 
     * @type {StudentStatus}
     * @memberof StudentBase
     */
    status?: StudentStatus;
}

export function StudentBaseFromJSON(json: any): StudentBase {
    return StudentBaseFromJSONTyped(json, false);
}

export function StudentBaseFromJSONTyped(json: any, ignoreDiscriminator: boolean): StudentBase {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': !exists(json, 'id') ? undefined : json['id'],
        'name': json['name'],
        'status': !exists(json, 'status') ? undefined : StudentStatusFromJSON(json['status']),
    };
}

export function StudentBaseToJSON(value?: StudentBase | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'status': StudentStatusToJSON(value.status),
    };
}